Fixed auth 401 error & incorrect session setting in fetcher

+++ client/src/components/fetcher.js
@@ -6,7 +6,13 @@ function Fetcher(){
     // Only required when using join call which requires game uuid
 
     this.get = function(action, payload){
+
+        console.log('fetching action: ', action);
+        console.log('with payload: ', payload);
+
         let token = sessionStorage.getItem('jwt');
+
+        console.log('and token: ', token);
 
         let body = {token: token};
 
@@ -14,7 +20,7 @@ function Fetcher(){
             body.gameID = payload
         }
 
-        fetch('http://'+domain+'8000/api/game/'+action,{
+        return fetch('http://'+domain+'8000/api/game/'+action,{
             method: 'POST',
             // mode: 'no-cors', // Only enable this for local debugging purposes
             body: JSON.stringify(body),
@@ -30,8 +36,10 @@ function Fetcher(){
         }).then((data)=>{
             console.log('data says: ', data);
 
-            sessionStorage.clear();
-            sessionStorage.setItem('jwt', data.token);
+            if(data.token !== undefined){
+                sessionStorage.clear();
+                sessionStorage.setItem('jwt', data.token);
+            }
 
             return data.token;
         });
